Move mustache template functions to separate class
Purpose of the FormattedString class is to support formatted strings as known from the C language. Hence
functions to replace mustache template by ValidHtml is moved to a new class MustachedTemplate.
